Move LevelButton component
